refactor(front_tests): move html query related functions to separate module
